fix: Rename message to block in some more places + Remove unused code
