+++ client/src/models/api/chrysalis/IMessageDetailsRequest.ts
